+++ src/day04.c
@@ -16,20 +16,20 @@ void d4_parse_input(char *input, size_t input_sz, StringView *d4_lines, int *d4_
 long long d4_count(StringView *d4_lines, int d4_ct, Point *d4_pts, int *d4_pts_head)
 {
     for (int l = 0; l < d4_ct; ++l) {
-        for (int c = 0; c < d4_lines[l].len; ++c) {
+        for (int c = 0; c < (int)d4_lines[l].len; ++c) {
             int nbors = 0;
             if (l > 0) {
                 if (d4_lines[l-1].buf[c] == '@') nbors++;
                 if (c > 0 && d4_lines[l-1].buf[c-1] == '@') nbors++;
-                if (c < d4_lines[l-1].len-1 && d4_lines[l-1].buf[c+1] == '@') nbors++;
+                if (c < (int)d4_lines[l-1].len-1 && d4_lines[l-1].buf[c+1] == '@') nbors++;
             }
             if (l < d4_ct-1) {
                 if (d4_lines[l+1].buf[c] == '@') nbors++;
                 if (c > 0 && d4_lines[l+1].buf[c-1] == '@') nbors++;
-                if (c < d4_lines[l+1].len-1 && d4_lines[l+1].buf[c+1] == '@') nbors++;
+                if (c < (int)d4_lines[l+1].len-1 && d4_lines[l+1].buf[c+1] == '@') nbors++;
             }
-            if (c < d4_lines[l].len-1 && d4_lines[l].buf[c+1] == '@') nbors++;
-            if (c > 0 && d4_lines[l].buf[c-1] == '@') nbors++;
+            if (c < (int)d4_lines[l].len-1 && d4_lines[l].buf[c+1] == '@') nbors++;
+            if (c > 0 && (int)d4_lines[l].buf[c-1] == '@') nbors++;
 
             if (nbors < 4 && d4_lines[l].buf[c] == '@') {
                 assert("d4_pts ran out of space" && (*d4_pts_head) + 1 < D4_PTS_MAX);
@@ -68,7 +68,7 @@ long long d4_solve_p2(char *input, size_t input_sz)
         do {
             Point p = d4_pts[--d4_pts_head];
             assert("d4_lines out of bounds" && D4_LINES_MAX > p.y);
-            assert("d4_lines out of bounds x" && d4_lines[p.y].len > p.x);
+            assert("d4_lines out of bounds x" && (int)d4_lines[p.y].len > p.x);
             d4_lines[p.y].buf[p.x] = '.';
         } while (d4_pts_head > 0);
     }
